Fix potential crash in BakeSkinning.
The crash would occur in _ExtendWorldTransformTimeSamples() if the
skinned prim or its parent had time-sampled transforms. Since multiple
skinning adapters can share the same skel adapter, this could result in
modifying the same list of time samples from multiple threads.

This change removes the parallel for loop over the skinning adapters,
but leaves some notes in the comments on the additional work that would
be required to achieve some parallelism.