Fix display of materials on instanceable prototype on point instancers.
The fix for #1449 looks like it broke material bindings in the case where
a point instancer prototype was marked as instanceable, but the material
was bound inside the prototype (on an instance proxy). The bound material
was being expressed with the instance proxy path instead of the prototype
path, which is how the material is actually identified in hydra. So now
the fix for #1449 is only applied to the instanceable prim, not the
contained instance proxy prims.

Fixes #1626