Get correct invalidations to coordSys sprims when their target USD prim is updated.
The issue is that the dirty bit used for an rprim transform and an sprim
coordSys transform are not the same. Within legacy scene delegate
emulation, dirty bits are converted to data source locators. When no
expected bits are set (for the sprim type), the change is suppressed.
Without emulation, the change would make it through but the wrong bit
would be set.

Also added tracking of animation or changes to camera-specific attributes
on coordsys prims (in case the renderer creates a full projection matrix
from a camera prim).