Fix a potential crash when updating skinned primitives.
With 6b3afc2f, skinned primitives are now replaced by an entry with
no type and no data source instead of being removed, since they have
child prims (the skinning computation Sprims).
When doing so, the entry's cache of computed primvars was never cleared.

This could result in a crash in scenarios such as switching to
displaying the result of a BakeSkinning operation. In this case, the
skinned primitive changes type from e.g. a Mesh (with a computed primvar
for the skinned positions) to empty, and then back to Mesh (which should
have no computed primvars). Since the entry's primvar cache was never cleared,
this incorrectly returns a stale list of computed primvars for the
updated mesh which later produces a crash.

When the prim type of an existing entry is being changed, the cached
primvars are now also cleared out while removing the old prim from the
render index.

Bug: #2809