Fix an issue where an unexpected time sample could be reported for skinning computations.
Previously, constant inputs to the skinning computation (e.g. the bind
transform) reported a single sample at a time offset of 0.
If none of the time-sampled computation inputs (e.g. joint transforms)
had a sample at that time, this would cause the computation to be
evaluated at an additional time sample (with the other computation
inputs being interpolated).

This could be confusing for users when configuring render options such
as the number of samples to use for motion blur, since this time sample
doesn't necessarily exist on the stage.

Since the boundaries of the sampling interval are always used when
sampling time-varying computation inputs, reporting the constant inputs
as having their time sample at the beginning of the interval avoids
potentially introducing an extra sample for the computation.