UsdImagingDrawModeSceneIndex: fixing scene index leaking out prim type and data source.
UsdImagingDrawModeSceneIndex::GetPrim no longer makes assumptions about how deep the prims by the draw mode standin go.

Draw mode standins: reimplementing _GetDescendantPrimType and _GetDescendantPrimSource to not returns prims that are not listed by _GetDescendantPaths.

Note that HdSceneIndex::GetPrim can be called with a path that is not in the scene index, for example, by the HdMergingSceneIndex. In this case, GetPrim has to return an empty prim (empty type and nullptr data source).

(Internal change: 2364208)