Improve performance of _RebuildInputsPathTable().
This became an n^2 loop when there were many inputs with non-overlapping scene
roots (e.g. native instancing with a large number of prototypes) and the
size of _inputsPathTable approached the number of inputs.

In this scenario the majority of the scene roots are of the form:
- /UsdNiPropagatedPrototypes/__hash_1__/__Prototype_1/UsdNiInstancer
...
- /UsdNiPropagatedPrototypes/__hash_N__/__Prototype_1/UsdNiInstancer

We can instead use FindSubtreeRange(sceneRoot) to directly find which
table entries each input should be added to, eliminating the HasPrefix()
check against every entry.

For a test scene with native instancing (17500 prototypes), this reduced
the overall time spent in _RebuildInputsPathTable() from around 5.5s to 18ms.