Improve performance of HdMergingSceneIndex::InsertInputScenes().
Modified for USD 21.11

This could produce n^2 behavior from querying and searching through
GetChildPrimPaths(parent) for each sibling path.

As an example, for native instancing with a large number of prototypes,
the input scenes' roots are of the form:
- /UsdNiPropagatedPrototypes/__hash_1__/__Prototype_1/UsdNiInstancer
...
- /UsdNiPropagatedPrototypes/__hash_N__/__Prototype_1/UsdNiInstancer

This results in N calls to GetChildPrimPaths("/UsdNiPropagatedPrototypes"),
which generates a list of N paths that is linearly searched through.

The change here is to make use the of the `visited` set to record
the existence of siblings that are discovered via GetChildPrimPaths(parent)

For a test scene with native instances with 17500 prototypes, this
reduces the time spent in this block from around 50s to 42ms